+++ FinApp/FinApp/appsettings.json
@@ -20,7 +20,7 @@
   },
   "AllowedHosts": "*",
   "Data": {
-    "DefaultConnection": "Data Source=(localdb)\\MSSQLLocalDb;Initial Catalog= FinApp;Trusted_Connection = true;MultipleActiveResultSets = true",
+    "DefaultConnection": "Data Source=(localdb)\\MSSQLLocalDB;Initial Catalog= FinApp;Trusted_Connection = true;MultipleActiveResultSets = true",
     "ConnectionString": "Server=192.168.195.147;Database=FinApp;User Id=Lv-429.Net;Password=Lv-429.Net"
   },
   "JwtIssuerOptions": {
